+++ libs/docusaurus-plugin-docgen/src/scripts/docgenParser.ts
@@ -1,4 +1,4 @@
-import { withCustomConfig } from 'react-docgen-typescript';
+import { withCompilerOptions } from 'react-docgen-typescript';
 import mapValues from 'lodash/mapValues';
 import omit from 'lodash/omit';
 import orderBy from 'lodash/orderBy';
@@ -57,7 +57,6 @@ function createTsProgramContext(tsconfigPath: string, filesToParse: string[]): T
       ...config.options,
       noEmit: true,
     },
-    projectReferences: config.projectReferences,
   });
 
   const checker = program.getTypeChecker();
@@ -615,7 +614,7 @@ export function docgenParser({
   }
 
   /** React docgen integration */
-  return withCustomConfig(params.tsconfigPath, {
+  return withCompilerOptions(tsCtx.program.getCompilerOptions(), {
     savePropValueAsString: true,
     shouldExtractValuesFromUnion: true,
     shouldExtractLiteralValuesFromEnum: true,
@@ -623,7 +622,7 @@ export function docgenParser({
     shouldIncludePropTagMap: true,
     shouldIncludeExpression: true,
   })
-    .parse(filesToParse)
+    .parseWithProgramProvider(filesToParse, () => tsCtx.program)
     .map((doc) => {
       const parentTypes: Record<string, string[]> = {};